+++ sdk/core/azure-core-amqp/src/amqp/message_sender.cpp
@@ -37,7 +37,7 @@ namespace Azure { namespace Core { namespace Amqp { namespace _internal {
 
   void MessageSender::Open(Context const& context) { m_impl->Open(context); }
   void MessageSender::Close() { m_impl->Close(); }
-  std::tuple<MessageSendStatus, Models::AmqpValue> MessageSender::Send(
+  std::tuple<MessageSendStatus, Models::_internal::AmqpError> MessageSender::Send(
       Models::AmqpMessage const& message,
       Context const& context)
   {
@@ -129,7 +129,7 @@ namespace Azure { namespace Core { namespace Amqp { namespace _detail {
 
       // Cache the error we received in the OnDetach notification so we can return it to the user
       // on the next send which fails.
-      m_savedMessageError = Models::_internal::AmqpErrorFactory::ToAmqp(error);
+      m_savedMessageError = error;
     });
   }
 
@@ -281,19 +281,22 @@ namespace Azure { namespace Core { namespace Amqp { namespace _detail {
     (void)context;
   }
 
-  std::tuple<_internal::MessageSendStatus, Models::AmqpValue> MessageSenderImpl::Send(
+  std::tuple<_internal::MessageSendStatus, Models::_internal::AmqpError> MessageSenderImpl::Send(
       Models::AmqpMessage const& message,
       Context const& context)
   {
-    Azure::Core::Amqp::Common::_internal::
-        AsyncOperationQueue<Azure::Core::Amqp::_internal::MessageSendStatus, Models::AmqpValue>
-            sendCompleteQueue;
+    Azure::Core::Amqp::Common::_internal::AsyncOperationQueue<
+        Azure::Core::Amqp::_internal::MessageSendStatus,
+        Models::_internal::AmqpError>
+        sendCompleteQueue;
 
     QueueSend(
         message,
         [&sendCompleteQueue, this](
             Azure::Core::Amqp::_internal::MessageSendStatus sendResult,
             Models::AmqpValue deliveryStatus) {
+          Models::_internal::AmqpError error;
+
           // If the send failed. then we need to return the error. If the send completed because
           // of an error, it's possible that the deliveryStatus provided is null. In that case,
           // we use the cached saved error because it is highly likely to be better than
@@ -302,16 +305,36 @@ namespace Azure { namespace Core { namespace Amqp { namespace _detail {
           {
             if (deliveryStatus.IsNull())
             {
-              deliveryStatus = m_savedMessageError;
+              error = m_savedMessageError;
+            }
+            else
+            {
+              if (deliveryStatus.GetType() != Models::AmqpValueType::List)
+              {
+                throw std::runtime_error("Delivery status is not a list");
+              }
+              auto deliveryStatusAsList{deliveryStatus.AsList()};
+              if (deliveryStatusAsList.size() != 1)
+              {
+                throw std::runtime_error("Delivery Status list is not of size 1");
+              }
+              Models::AmqpValue firstState{deliveryStatusAsList[0]};
+              ERROR_HANDLE errorHandle;
+              if (!amqpvalue_get_error(firstState, &errorHandle))
+              {
+                Models::_internal::UniqueAmqpErrorHandle uniqueError{
+                    errorHandle}; // This will free the error handle when it goes out of scope.
+                error = Models::_internal::AmqpErrorFactory::FromUamqp(errorHandle);
+              }
             }
           }
           else
           {
             // If we successfully sent the message, then whatever saved error should be cleared,
             // it's no longer valid.
-            m_savedMessageError = Models::AmqpValue();
+            m_savedMessageError = Models::_internal::AmqpError();
           }
-          sendCompleteQueue.CompleteOperation(sendResult, deliveryStatus);
+          sendCompleteQueue.CompleteOperation(sendResult, error);
         },
         context);
     auto result = sendCompleteQueue.WaitForPolledResult(context, *m_session->GetConnection());