+++ sdk/core/azure-core-amqp/src/amqp/session.cpp
@@ -238,8 +238,6 @@ namespace Azure { namespace Core { namespace Amqp { namespace _detail {
       m_claimsBasedSecurity = std::make_shared<ClaimsBasedSecurityImpl>(shared_from_this());
     }
     auto accessToken = GetConnection()->GetSecurityToken(audience, context);
-    Log::Stream(Logger::Level::Informational)
-        << "Authenticate with audience: " << audience << ", token: " << accessToken;
 
     m_claimsBasedSecurity->SetTrace(GetConnection()->EnableTrace());
     if (!m_cbsOpen)